chore!: Drop support for Python 3.9
BREAKING CHANGE: nixpkgs unstable now has a click version (8.2.x) which
drops support for Python 3.9 *and* breaks my tests. I don't think it's
possible to fix those tests in a way which will work both with click
before and after 8.2.x). So basically I need to drop Python 3.9 in
order to upgrade click in order to fix build in nixpkgs.
Python 3.9 is EOL though so I guess it's fine.